+++ server/src/ascot/index.ts
@@ -11,16 +11,8 @@ import {
 	SymbolTag,
 } from "vscode-languageserver";
 import { connection } from "../utils/variables";
-import {
-	getServerSpec,
-	makeRESTRequest,
-	buildMemberMetadataQuery,
-	buildSuperclassesQuery,
-	buildClassTypeQuery,
-	resolveStubbedMethod,
-	MemberMetadataRow,
-	documaticHtmlToMarkdown,
-} from "../utils/functions";
+import { getServerSpec, makeRESTRequest, documaticHtmlToMarkdown } from "../utils/functions";
+import { ServerSpec, QueryData } from "../utils/types";
 import { URI } from "vscode-uri";
 
 const libDir = path.resolve(__dirname, "../lib");
@@ -32,6 +24,97 @@ function getCoreModule(name: string): WebAssembly.Module {
 const ORIGIN: wit.Position = { line: 0, character: 0 };
 const ZERO_RANGE: wit.Range = { start: ORIGIN, end: ORIGIN };
 const CACHE_TTL_MS = 60 * 60 * 1000;
+
+interface MemberMetadataRow {
+	MemberType: "method" | "property" | "parameter";
+	Description: string;
+	FormalSpec: string;
+	ReturnType: string;
+	Stub: string;
+	// Bit columns come back as "0"/"1" strings from a plain query, but a UNION ALL
+	// with a numeric literal column coerces them to 0/1 numbers. Accept either.
+	ClassMethod: "0" | "1" | 0 | 1;
+	Deprecated: "0" | "1" | 0 | 1;
+}
+
+/** Which %Dictionary.Compiled* table(s) to search for a member. "auto" mirrors the
+ * hover fallback (a %SYSTEM class is method-only, else method-or-property); "any"
+ * also searches parameters, for callers that don't know the kind up front. */
+type MemberMetadataKind = "parameter" | "method" | "property" | "auto" | "any";
+
+/** Build the /action/query request for a class member's metadata. The SELECT column
+ * list is identical across kinds so callers can read a uniform MemberMetadataRow. */
+function buildMemberMetadataQuery(baseclass: string, unquotedname: string, kind: MemberMetadataKind): QueryData {
+	const method =
+		"SELECT 'method' AS MemberType, Description, FormalSpec, ReturnType, Stub, ClassMethod, Deprecated " +
+		"FROM %Dictionary.CompiledMethod WHERE Parent = ? AND Name = ?";
+	const property =
+		"SELECT 'property' AS MemberType, Description, NULL AS FormalSpec, " +
+		"CASE WHEN Collection IS NOT NULL THEN Collection||' Of '||Type ELSE Type END AS ReturnType, " +
+		"NULL AS Stub, 0 AS ClassMethod, Deprecated " +
+		"FROM %Dictionary.CompiledProperty WHERE Parent = ? AND (Name = ? OR ? %INLIST $LISTFROMSTRING($TRANSLATE(Aliases,' ')))";
+	const parameter =
+		"SELECT 'parameter' AS MemberType, Description, NULL AS FormalSpec, Type AS ReturnType, " +
+		"NULL AS Stub, 0 AS ClassMethod, Deprecated FROM %Dictionary.CompiledParameter WHERE Parent = ? AND Name = ?";
+	const nospace = unquotedname.replace(/\s+/g, "");
+	switch (kind) {
+		case "method":
+			return { query: method, parameters: [baseclass, unquotedname] };
+		case "parameter":
+			return { query: parameter, parameters: [baseclass, unquotedname] };
+		case "property":
+			return { query: property, parameters: [baseclass, unquotedname, nospace] };
+		case "auto":
+			if (baseclass.startsWith("%SYSTEM")) {
+				return { query: method, parameters: [baseclass, unquotedname] };
+			}
+			return {
+				query: `${method} UNION ALL ${property}`,
+				parameters: [baseclass, unquotedname, baseclass, unquotedname, nospace],
+			};
+		case "any":
+			return {
+				query: `${method} UNION ALL ${property} UNION ALL ${parameter}`,
+				parameters: [baseclass, unquotedname, baseclass, unquotedname, nospace, baseclass, unquotedname],
+			};
+	}
+}
+
+/** Build the /action/query request for a class's direct superclasses. `Super` is a
+ * comma-separated list in declaration order (nearest first). */
+function buildSuperclassesQuery(cls: string): QueryData {
+	return {
+		query: "SELECT Super FROM %Dictionary.CompiledClass WHERE Name = ?",
+		parameters: [cls],
+	};
+}
+
+function buildClassTypeQuery(cls: string): QueryData {
+	return {
+		query: "SELECT ClassType FROM %Dictionary.CompiledClass WHERE Name = ?",
+		parameters: [cls],
+	};
+}
+
+/** A method with a non-empty Stub is generated by member inheritance; its real
+ * metadata lives in a subtable keyed by the Stub's segments (i/q/a/n). */
+async function resolveStubbedMethod(
+	server: ServerSpec,
+	baseclass: string,
+	stub: string,
+): Promise<MemberMetadataRow | undefined> {
+	const [origin, subname, kindchar] = stub.split(".");
+	const table = { i: "Index", q: "Query", a: "Property", n: "Constraint" }[kindchar];
+	if (!table) return undefined;
+	const respdata = await makeRESTRequest("POST", 1, "/action/query", server, {
+		query:
+			`SELECT 'method' AS MemberType, Description, FormalSpec, ReturnType, '' AS Stub, ClassMethod, Deprecated ` +
+			`FROM %Dictionary.Compiled${table}Method WHERE Name = ? AND parent->Parent = ? AND parent->Name = ?`,
+		parameters: [subname, baseclass, origin],
+	});
+	const rows = respdata?.data?.result?.content;
+	return Array.isArray(rows) && rows.length > 0 ? rows[0] : undefined;
+}
 
 class IrisConnection {
 	constructor(
